type checking, ruff, pre-commit passing
