+++ .gitignore
@@ -294,3 +294,4 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+.aider*
