--- tap_service_titan/schemas/__init__.py
@@ -1,1 +1,0 @@
-"""JSON schema files for the REST API."""
