+++ tap_service_titan/streams/__init__.py
@@ -1,0 +1,52 @@
+"""All streams for the ServiceTitan tap."""
+
+from . import (
+    accounting,
+    crm,
+    customer_interactions,
+    dispatch,
+    equipment_systems,
+    forms,
+    inventory,
+    job_booking,
+    jpm,
+    marketing,
+    marketing_ads,
+    marketing_reputation,
+    memberships,
+    payroll,
+    pricebook,
+    reporting,
+    sales_and_estimates,
+    scheduling_pro,
+    service_agreements,
+    settings,
+    task_management,
+    telecom,
+)
+
+__all__ = [
+    "accounting",
+    "crm",
+    "customer_interactions",
+    "dispatch",
+    "equipment_systems",
+    "forms",
+    "inventory",
+    "job_booking",
+    "jpm",
+    "marketing",
+    "marketing_ads",
+    "marketing_reputation",
+    "memberships",
+    "payroll",
+    "pricebook",
+    "reporting",
+    "sales_and_estimates",
+    "scheduling_pro",
+    "service_agreements",
+    "settings",
+    "streams",
+    "task_management",
+    "telecom",
+]
