+++ tap_service_titan/streams/accounting.py
@@ -1,0 +1,434 @@
+"""Accounting streams for the ServiceTitan tap."""
+
+from __future__ import annotations
+
+import typing as t
+from functools import cached_property
+
+from singer_sdk import typing as th  # JSON Schema typing helpers
+
+from tap_service_titan.client import (
+    ServiceTitanExportStream,
+    ServiceTitanStream,
+)
+
+
+class InvoicesStream(ServiceTitanExportStream):
+    """Define invoices stream."""
+
+    name = "invoices"
+    primary_keys: t.ClassVar[list[str]] = ["id"]
+    replication_key: str = "modifiedOn"
+
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property("syncStatus", th.StringType),
+        th.Property("summary", th.StringType),
+        th.Property("referenceNumber", th.StringType),
+        th.Property("invoiceDate", th.StringType),
+        th.Property("dueDate", th.StringType),
+        th.Property("subTotal", th.StringType),
+        th.Property("salesTax", th.StringType),
+        th.Property(
+            "salesTaxCode",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("taxRate", th.NumberType),
+            ),
+        ),
+        th.Property("total", th.StringType),
+        th.Property("balance", th.StringType),
+        th.Property(
+            "invoiceType",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "customer",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "customerAddress",
+            th.ObjectType(
+                th.Property("street", th.StringType),
+                th.Property("unit", th.StringType),
+                th.Property("city", th.StringType),
+                th.Property("state", th.StringType),
+                th.Property("zip", th.StringType),
+                th.Property("country", th.StringType),
+            ),
+        ),
+        th.Property(
+            "location",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "locationAddress",
+            th.ObjectType(
+                th.Property("street", th.StringType),
+                th.Property("unit", th.StringType),
+                th.Property("city", th.StringType),
+                th.Property("state", th.StringType),
+                th.Property("zip", th.StringType),
+                th.Property("country", th.StringType),
+            ),
+        ),
+        th.Property(
+            "businessUnit",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "batch",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("number", th.StringType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property("depositedOn", th.StringType),
+        th.Property("createdOn", th.StringType),
+        th.Property("modifiedOn", th.DateTimeType),
+        th.Property("adjustmentToId", th.IntegerType),
+        th.Property(
+            "job",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("number", th.StringType),
+                th.Property("type", th.StringType),
+            ),
+        ),
+        th.Property("projectId", th.IntegerType),
+        th.Property(
+            "royalty",
+            th.ObjectType(
+                th.Property("status", th.StringType),
+                th.Property("date", th.StringType),
+                th.Property("sentOn", th.StringType),
+                th.Property("memo", th.StringType),
+            ),
+        ),
+        th.Property(
+            "employeeInfo",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("modifiedOn", th.DateTimeType),
+            ),
+        ),
+        th.Property("commissionEligibilityDate", th.StringType),
+        th.Property("sentStatus", th.StringType),  # Assuming this is a string type
+        th.Property("reviewStatus", th.StringType),  # Assuming this is a string type
+        th.Property(
+            "assignedTo",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "items",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("id", th.IntegerType),
+                    th.Property("description", th.StringType),
+                    th.Property("quantity", th.StringType),
+                    th.Property("cost", th.StringType),
+                    th.Property("totalCost", th.StringType),
+                    th.Property("inventoryLocation", th.StringType),
+                    th.Property("price", th.StringType),
+                    th.Property(
+                        "type", th.StringType
+                    ),  # Assuming this is a string type
+                    th.Property("skuName", th.StringType),
+                    th.Property("skuId", th.IntegerType),
+                    th.Property("total", th.StringType),
+                    th.Property("inventory", th.BooleanType),
+                    th.Property("taxable", th.BooleanType),
+                    th.Property(
+                        "generalLedgerAccount",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                            th.Property("number", th.StringType),
+                            th.Property("type", th.StringType),
+                            th.Property("detailType", th.StringType),
+                        ),
+                    ),
+                    th.Property(
+                        "costOfSaleAccount",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                            th.Property("number", th.StringType),
+                            th.Property("type", th.StringType),
+                            th.Property("detailType", th.StringType),
+                        ),
+                    ),
+                    th.Property(
+                        "assetAccount",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                            th.Property("number", th.StringType),
+                            th.Property("type", th.StringType),
+                            th.Property("detailType", th.StringType),
+                        ),
+                    ),
+                    th.Property("membershipTypeId", th.IntegerType),
+                    th.Property(
+                        "itemGroup",
+                        th.ObjectType(
+                            th.Property("rootId", th.IntegerType),
+                            th.Property("name", th.StringType),
+                        ),
+                    ),
+                    th.Property("displayName", th.StringType),
+                    th.Property("soldHours", th.NumberType),
+                    th.Property("modifiedOn", th.DateTimeType),
+                    th.Property("serviceDate", th.StringType),
+                    th.Property("order", th.IntegerType),
+                    th.Property(
+                        "businessUnit",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                        ),
+                    ),
+                )
+            ),
+        ),
+        th.Property(
+            "customFields",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("name", th.StringType),
+                    th.Property("value", th.StringType),
+                )
+            ),
+        ),
+        th.Property("active", th.BooleanType),
+    ).to_dict()
+
+    @cached_property
+    def path(self) -> str:
+        """Return the API path for the stream."""
+        return f"/accounting/v2/tenant/{self._tap.config['tenant_id']}/export/invoices"
+
+
+class InvoiceItemsStream(ServiceTitanExportStream):
+    """Define invoice items stream."""
+
+    name = "invoice_items"
+    primary_keys: t.ClassVar[list[str]] = ["id"]
+    replication_key: str = "modifiedOn"
+
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property("description", th.StringType),
+        th.Property("quantity", th.StringType),
+        th.Property("cost", th.StringType),
+        th.Property("totalCost", th.StringType),
+        th.Property("inventoryLocation", th.StringType),
+        th.Property("price", th.StringType),
+        th.Property("type", th.StringType),
+        th.Property("skuName", th.StringType),
+        th.Property("skuId", th.IntegerType),
+        th.Property("total", th.StringType),
+        th.Property("inventory", th.BooleanType),
+        th.Property("taxable", th.BooleanType),
+        th.Property(
+            "generalLedgerAccount",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("number", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("detailType", th.StringType),
+            ),
+        ),
+        th.Property(
+            "costOfSaleAccount",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("number", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("detailType", th.StringType),
+            ),
+        ),
+        th.Property(
+            "assetAccount",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("number", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("detailType", th.StringType),
+            ),
+        ),
+        th.Property("membershipTypeId", th.IntegerType),
+        th.Property(
+            "itemGroup",
+            th.ObjectType(
+                th.Property("rootId", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property("displayName", th.StringType),
+        th.Property("soldHours", th.NumberType),
+        th.Property("modifiedOn", th.DateTimeType),
+        th.Property("serviceDate", th.DateTimeType),
+        th.Property("order", th.IntegerType),
+        th.Property(
+            "businessUnit",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property("active", th.BooleanType),
+        th.Property("invoiceId", th.IntegerType),
+    ).to_dict()
+
+    @cached_property
+    def path(self) -> str:
+        """Return the API path for the stream."""
+        return f"/accounting/v2/tenant/{self._tap.config['tenant_id']}/export/invoice-items"  # noqa: E501
+
+
+class EstimateItemsStream(ServiceTitanStream):
+    """Define estimate items stream."""
+
+    name = "estimate_items"
+    primary_keys: t.ClassVar[list[str]] = ["id"]
+    replication_key: str = "modifiedOn"
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property(
+            "sku",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("displayName", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("soldHours", th.NumberType),
+                th.Property("generalLedgerAccountId", th.IntegerType),
+                th.Property("generalLedgerAccountName", th.StringType),
+                th.Property("modifiedOn", th.DateTimeType),
+            ),
+        ),
+        th.Property("skuAccount", th.StringType),
+        th.Property("description", th.StringType),
+        th.Property("membershipTypeId", th.IntegerType),
+        th.Property("qty", th.NumberType),
+        th.Property("unitRate", th.NumberType),
+        th.Property("total", th.NumberType),
+        th.Property("unitCost", th.NumberType),
+        th.Property("totalCost", th.NumberType),
+        th.Property("itemGroupName", th.StringType),
+        th.Property("itemGroupRootId", th.IntegerType),
+        th.Property("createdOn", th.DateTimeType),
+        th.Property(
+            "modifiedOn", th.DateTimeType
+        ),  # Assuming datetime format as string
+        th.Property("chargeable", th.BooleanType),
+    ).to_dict()
+
+    @cached_property
+    def path(self) -> str:
+        """Return the API path for the stream."""
+        return f"/sales/v2/tenant/{self._tap.config['tenant_id']}/estimates/items"
+
+
+class PaymentsStream(ServiceTitanExportStream):
+    """Define payments stream."""
+
+    name = "payments"
+    primary_keys: t.ClassVar[list[str]] = ["id"]
+    replication_key: str = "modifiedOn"
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property("syncStatus", th.StringType),
+        th.Property("referenceNumber", th.StringType),
+        th.Property("date", th.DateTimeType),
+        th.Property("type", th.StringType),
+        th.Property("typeId", th.StringType),
+        th.Property("total", th.StringType),
+        th.Property("unappliedAmount", th.StringType),
+        th.Property("memo", th.StringType),
+        th.Property(
+            "customer",
+            th.ObjectType(
+                th.Property("id", th.IntegerType), th.Property("name", th.StringType)
+            ),
+        ),
+        th.Property(
+            "businessUnit",
+            th.ObjectType(
+                th.Property("id", th.IntegerType), th.Property("name", th.StringType)
+            ),
+        ),
+        th.Property(
+            "batch",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("number", th.StringType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property("createdBy", th.StringType),
+        th.Property(
+            "generalLedgerAccount",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("number", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("detailType", th.StringType),
+            ),
+        ),
+        th.Property(
+            "appliedTo",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("appliedId", th.IntegerType),
+                    th.Property("appliedTo", th.IntegerType),
+                    th.Property("appliedAmount", th.StringType),
+                    th.Property("appliedOn", th.DateTimeType),
+                    th.Property("appliedBy", th.StringType),
+                    th.Property("appliedToReferenceNumber", th.StringType),
+                )
+            ),
+        ),
+        th.Property(
+            "customFields",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("name", th.StringType),
+                    th.Property("value", th.StringType),
+                )
+            ),
+        ),
+        th.Property("authCode", th.StringType),
+        th.Property("checkNumber", th.StringType),
+        th.Property("modifiedOn", th.DateTimeType),
+        th.Property("createdOn", th.DateTimeType),
+        th.Property("active", th.BooleanType),
+    ).to_dict()
+
+    @cached_property
+    def path(self) -> str:
+        """Return the API path for the stream."""
+        return f"/accounting/v2/tenant/{self._tap.config['tenant_id']}/export/payments"
