+++ tap_service_titan/streams/customer_interactions.py
@@ -1,0 +1,1 @@
+"""Customer interactions streams for the ServiceTitan tap."""
