+++ tap_service_titan/streams/equipment_systems.py
@@ -1,0 +1,1 @@
+"""Equipment systems streams for the ServiceTitan tap."""
