+++ tap_service_titan/streams/job_booking.py
@@ -1,0 +1,1 @@
+"""Job booking streams for the ServiceTitan tap."""
