+++ tap_service_titan/streams/marketing_ads.py
@@ -1,0 +1,1 @@
+"""Marketing ads streams for the ServiceTitan tap."""
