+++ tap_service_titan/streams/payroll.py
@@ -1,0 +1,1 @@
+"""Payroll streams for the ServiceTitan tap."""
