+++ tap_service_titan/streams/pricebook.py
@@ -1,0 +1,1 @@
+"""Pricebook streams for the ServiceTitan tap."""
