+++ tap_service_titan/streams/scheduling_pro.py
@@ -1,0 +1,1 @@
+"""Scheduling streams for the ServiceTitan tap."""
